// src/app/core/models/settings/radiologist-signup/GblDistrict.ts
export class GblDistrict {
    DistrictId?: number;
    DistrictCode!: string;
    DistrictName!: string;
    CountryId?: number;
    ProvinceId?: number;
    GeoId?: number;
    CreatedBy?: number;
    CreatedOn?: Date;
    LastModifiedBy?: number;
    LastModifiedOn?: Date;
    IsActive?: boolean;
}