// src/app/core/models/settings/radiologist-signup/GblSubdistrict.ts
export class GblSubdistrict {
    SubdistrictId!: number;
    SubdistrictCode!: string;
    SubdistrictName!: string;
    CountryId?: number;
    DistrictId?: number ;
    GeoId?: number;
    CreatedBy?: number;
    CreatedOn?: Date ;
    LastModifiedBy?: number;
    LastModifiedOn?: Date;
    IsActive?: boolean;
}